chore(frontend): move DatabaseGroup/utils to React layer
Last single-file framework-neutral helper dir under components/.
1 React caller, 0 Vue callers. The src/components/DatabaseGroup/
directory is removed entirely.

The moved file still has two cross-framework type imports
(ExprEditor/context's OptionConfig, RemoteResourceSelector/types'
ResourceSelectOption) — same situation as SensitiveData/components-utils
landed in the previous commit. Both can't be cleaned until ExprEditor/
context migrates.

Extended no-react-to-vue-imports.test.ts ban list to include
@/components/DatabaseGroup.
